+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,16 @@
 # playground
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @chatool/styles@0.1.0
+  - @chatool/utils@0.1.0
+  - @chatool/ui@0.2.0
+  - @chatool/icons@0.2.0
+  - @chatool/api@0.1.0
+
 ## 0.0.1
 
 ### Patch Changes